INE5404: Adicionada versão 0.7
Por algum motivo, eu achei que seria uma boa ideia
começar a numeração na versão 0.7.

+++ INE5404/src/visao/AdaptadorImagem.java
@@ -6,7 +6,7 @@
 
 package visao;
 
-import modelo.estruturasDeDados.Coordenada;
+import modelo.Coordenada;
 import edugraf.jadix.fachada.ComponenteDix;
 import edugraf.jadix.fachada.PaginaDix;
 import edugraf.jadix.fachada.TiposDeComponentesDix;
@@ -29,6 +29,10 @@ public class AdaptadorImagem implements TipoImagem {
      *            Posição inicial da imagem
      * @param fonte
      *            fonte do arquivo que será mostrado na tela.
+     * @param largura
+     *            Largura da imagem.
+     * @param altura
+     *            Altura da imagem.
      */
     public AdaptadorImagem(PaginaDix tela, Coordenada posicao, String fonte,
             int largura, int altura) {
@@ -40,8 +44,9 @@ public class AdaptadorImagem implements TipoImagem {
 
     @Override
     public AdaptadorImagem centrarSeEm(Coordenada posicao) {
-        imagem.fixarCoordenadas(new Coordenadas(posicao.x
-                - imagem.obterLargura() / 2, posicao.y - imagem.obterAltura() / 2 ));
+        imagem.fixarCoordenadas(new Coordenadas((int) (posicao.x - imagem
+                .obterLargura() / 2),
+                (int) (posicao.y - imagem.obterAltura() / 2)));
         return this;
     }
 
@@ -60,7 +65,7 @@ public class AdaptadorImagem implements TipoImagem {
      * @return Um objeto Coordenadas equivalente a <u>c</u>
      */
     protected static final Coordenadas converterCoordenada(Coordenada c) {
-        return new Coordenadas(c.x, c.y);
+        return new Coordenadas((int) c.x, (int) c.y);
     }
 
     /**
